requirements.txt
requests>=2.25.1
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.5.0 